[FIX] : move blogPopover to component
