+++ components/adminComp/workspacesCompTable/index.tsx
@@ -8,7 +8,7 @@ import { blogTableType } from '@/libs/types/blogTableType';
 import { Tag } from '@/libs/types/tagType';
 import { WorkspacesResponseWithPagination, workspacesType } from '@/libs/types/workspacesType';
 import { convertIsoToDateTime, transformBlogTableType } from '@/libs/utils';
-import BlogPopover from '@/pages/blogs/myBlogs/blogPopover';
+import BlogPopover from '@/components/blogPopover';
 import {
   Flex,
   useMantineTheme,
